pyproject.toml
[project]
name = "ai-chat-history"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "python-dotenv>=1.2.1",
]
